LPS-102957 Move all views and related classes to the new app (fragment collection and fragment entries)
