+++ modules/apps/layout/layout-page-template-admin-web/src/main/java/com/liferay/layout/page/template/admin/web/internal/servlet/taglib/clay/LayoutPrototypeVerticalCard.java
@@ -19,7 +19,7 @@ import com.liferay.frontend.taglib.clay.servlet.taglib.soy.VerticalCard;
 import com.liferay.frontend.taglib.clay.servlet.taglib.util.DropdownItem;
 import com.liferay.frontend.taglib.clay.servlet.taglib.util.LabelItem;
 import com.liferay.frontend.taglib.clay.servlet.taglib.util.LabelItemList;
-import com.liferay.layout.page.template.admin.web.internal.constants.LayoutTemplateAdminWebKeys;
+import com.liferay.layout.page.template.admin.web.internal.constants.LayoutPageTemplateAdminWebKeys;
 import com.liferay.layout.page.template.admin.web.internal.servlet.taglib.util.LayoutPrototypeActionDropdownItemsProvider;
 import com.liferay.portal.kernel.dao.search.RowChecker;
 import com.liferay.portal.kernel.language.LanguageUtil;
@@ -80,7 +80,7 @@ public class LayoutPrototypeVerticalCard
 
 	@Override
 	public String getDefaultEventHandler() {
-		return LayoutTemplateAdminWebKeys.
+		return LayoutPageTemplateAdminWebKeys.
 			LAYOUT_PROTOTYPE_DROPDOWN_DEFAULT_EVENT_HANDLER;
 	}
