+++ modules/apps/layout/layout-page-template-admin-web/src/main/resources/META-INF/resources/edit_layout_page_template_collection.jsp
@@ -29,8 +29,8 @@ portletDisplay.setURLBack(redirect);
 renderResponse.setTitle((layoutPageTemplateCollection != null) ? layoutPageTemplateCollection.getName() : LanguageUtil.get(request, "add-collection"));
 %>
 
-<portlet:actionURL name="/layout/edit_layout_page_template_collection" var="editLayoutPageTemplateCollectionURL">
-	<portlet:param name="mvcRenderCommandName" value="/layout/edit_layout_page_template_collection" />
+<portlet:actionURL name="/layout_page_template/edit_layout_page_template_collection" var="editLayoutPageTemplateCollectionURL">
+	<portlet:param name="mvcRenderCommandName" value="/layout_page_template/edit_layout_page_template_collection" />
 	<portlet:param name="tabs1" value="page-templates" />
 </portlet:actionURL>
